Hopefully produce better diff
